pos correction works
We need the make separate teleops for red and blue alliance bc the straitening out parallel to the field wall closest would be different depending on what alliance we will be on. In addition, we will want to not reset the auto in teleop and only do it in auto if that is possible in order to eliminate having to spend time at the beginning of teleop each time resetting the imu.